+++ projects/frontend/es-app/src/guis/menus/GuiMenuAdvancedSettings.cpp
@@ -11,6 +11,7 @@
 #include "GuiMenuSystemList.h"
 #include "GuiMenuKodiSettings.h"
 #include "GuiMenuCRT.h"
+#include "GuiMenuResolutionSettings.h"
 #include <guis/MenuMessages.h>
 #include <utils/locale/LocaleHelper.h>
 //#include <components/OptionListComponent.h>
@@ -44,11 +45,15 @@ GuiMenuAdvancedSettings::GuiMenuAdvancedSettings(WindowManager& window, SystemMa
   if (Board::Instance().CanHaveCRTBoard())
     AddSubMenu(_("RECALBOX CRT"), (int)Components::CrtSubMenu, _(MENUMESSAGE_ADVANCED_CRT_HELP_MSG));
 
-  // Adult games
-  AddSwitch(_("HIDE ADULT GAMES IN ALL SYSTEMS"), RecalboxConf::Instance().GetFilterAdultGames(), (int)Components::AdultGames, this, _(MENUMESSAGE_GAMELISTOPTION_HIDE_ADULT_MSG));
+  // RESOLUTION
+  if (Renderer::GetResolutionList().size() > 1 && Board::Instance().CrtBoard().GetCrtAdapter() == CrtAdapterType::None)
+    AddSubMenu(_("RESOLUTIONS"), (int)Components::ResolutionSubMenu, _(MENUMESSAGE_ADVANCED_RESOLUTION_HELP_MSG));
 
   // Custom config for systems
   AddSubMenu(_("ADVANCED EMULATOR CONFIGURATION"), (int)Components::AdvancedSubMenu, _(MENUMESSAGE_ADVANCED_EMULATOR_ADVANCED_HELP_MSG));
+
+  // Adult games
+  AddSwitch(_("HIDE ADULT GAMES IN ALL SYSTEMS"), RecalboxConf::Instance().GetFilterAdultGames(), (int)Components::AdultGames, this, _(MENUMESSAGE_GAMELISTOPTION_HIDE_ADULT_MSG));
 
   //Kodi
   if (RecalboxSystem::kodiExists())
@@ -223,6 +228,7 @@ void GuiMenuAdvancedSettings::SwitchComponentChanged(int id, bool status)
     case Components::SecuritySubMenu:
     case Components::FactoryReset:
     case Components::CrtSubMenu:
+    case Components::ResolutionSubMenu:
     default: break;
   }
 }
@@ -236,6 +242,7 @@ void GuiMenuAdvancedSettings::SubMenuSelected(int id)
     case Components::VirtualSubMenu: mWindow.pushGui(new GuiMenuVirtualSystems(mWindow, mSystemManager)); break;
     case Components::AdvancedSubMenu: mWindow.pushGui(new GuiMenuSystemList(mWindow, mSystemManager)); break;
     case Components::KodiSubMenu: mWindow.pushGui(new GuiMenuKodiSettings(mWindow)); break;
+    case Components::ResolutionSubMenu: mWindow.pushGui(new GuiMenuResolutionSettings(mWindow, mSystemManager)); break;
     case Components::FactoryReset: ResetFactory(); break;
     case Components::OverclockList:
     case Components::AdultGames: