+++ RELEASE-NOTES.md
@@ -22,6 +22,7 @@ as release notes for end user on a Recalbox upgrade.
 - Add Pi400 power button support (Fn + F10 to turn off)
 - Add game regions flag in gamelist
 - Make default no image in gamelist themable
+- Add resolution menu in Advanced Settings > Resolutions
 
 ### Improvements